Rename timer.retry to timer.heartbeat
